Create 542_01_matrix.py
Solution Summary
The solution utilizes BFS by collecting all the cells that are 0, and setting all that are equal to 1, to infinitely. We then use BFS and check if any neighboring cell is greater in value compared to the current cell, if so we can update the neighbor cell using the current cell's value + 1. This will effectively the distance to the nearest 0 given any cell, because cells that are 0 will remain 0, and cells that are not will be updated to be greater than 0.


Loom Link
https://www.loom.com/share/ff6f70f5475f4789902b65481be2e81c