Create 735_asteroid_collision.py
Solution Summary
The solution uses a stack to handle collisions, three cases in particular. If a collision is about to occur we pop from the stack when the incoming asteroid is either >= to the asteroid at the top of the stack, where we handle the continuation/breakage of the loop depending on the condition. Lastly, we can simulate the incoming-smaller asteroid being destroyed by simply breaking out of the while condition. 

Loom Link
https://www.loom.com/share/da9fd2c19e3a428db7ce41b9ef867e45