Create 659_design_hit_counter.py
Summary Solution

The solution uses a list to store hits, because hits are in increasing order this means the list will always be in sorted order. Thus, to extract the number of hits within the last 300 seconds at we have to do is implement binary search to find "starting" position (lo) of the hits to which we can then subtract from the length of self.times. This will pretty much be the slice of the list that contain the hits (as an integer).