Create 130_surrounded_regions.py
Solution Summary
The solution uses DFS to convert any border cells to a temporary value such that they can be converted back to O's due to the problem constraint. Once we have converted the border-cells we then convert any remaining O's within the board because we know they must be surrounded, if any of these O's were neighboring a "T" cell we know know they cant be fully be surrounded by X's. Lastly, we convert any of the T (bordering-cells) back to O's during the conversion of the potential inner "O" cells.

Loom Link
https://www.loom.com/share/c689eb656ef94e7da9a04f5fd498cd36