Create 208_implement_trie_prefix_tree.py
Summary
The solution makes use of additional Node class to create the Trie tree. Each node contains a dictionary that can store characters a-z at each level. We set the isWord flag to true on the final letter of the input "word" when we insert into the tree. 


Loom Link
https://www.loom.com/share/c7591207185f482cad48f119fdfee3ee