Create 215_k_largest_element_in_an_array.py
Summary
The solution implements a heap, a structure that holds the minimum value at the 0th index. We fill our heap until it reaches size k and continuously pop the smallest element in the heap when we run into a larger one. This means the 0th element will always be the Kth largest in the original array.

Loom Link
https://www.loom.com/share/e8511247fc5d4da69ad71a283459fbdf