Create 55_jump_game.py
Solution Summary
The solution makes use of a variable, "moves", to track the number in which we can advance in the main loop. In the main loop, we check if the "moves" is negative, signifying that we can no longer progress in the loop, thus we return False. We change the value of "moves" every time we run into a larger value (maximum moves at this specific index). In either event we must decrement by one as we progress in the loop.

Loom Link
https://www.loom.com/share/5505e8e6e1b745be86c035c54672b8a6