Create 56_merge_intervals.py
Solution Summary
The solution makes use of sorting the intervals to easily check for overlaps. We merge under the condition that the end of the interval at the top of our stack is less than or equal to the current interval's start. This will update the top of our stack to get the maximum between the two end points to resolve any overlapping.

Loom Link
https://www.loom.com/share/7eb6c35ec9c54bc7aa53e67ca0fbc3c5