Create 23_merge_k_sorted_lists.py
Solution Summary
This function makes use of a heap to first heappush every element in ever list within lists. We then construct a node "cur" which will be used to construct the remainder of the list. Lastly we pop from our pq using heappop to create the next Node, a new ListNode and iterate to it.

Loom Link