Create 153_find_minimum_in_rotated_sorted_array.py
Solution Summary
The solution makes use of Binary Search to achieve the constraint of O(log N) runtime. We alter the search by checking if the element at "mid" is less than the element at "hi", if so we know that the search should be conducted from "lo" to "mid" (the elements lesser in value). If this condition fails, we know that the element at "mid" is greater than the element at hi, thus the section of elements of lesser value reside from "mid+1" to "hi".


Loom Link
https://www.loom.com/share/c75664b7dae14be9b88c799514085e4e