Create 207_course_schedule.py
Solution Summary
The solution uses DFS to detect any potential cycle within the prerequisite's list. We construct an adjacency list and visited list to alter and track the state of each course. We only need one false return, meaning a there a cycle to which we cannot complete all courses.  

Loom Link
https://www.loom.com/share/7c14bed6cba243afa6e96b9df5ee8418