Create 417_pacific_atlantic_waterflow.py
Solution Summary
The solution uses BFS to find cells that can reach both the Pacific and Atlantic oceans. We initialize the search from the cells that are on the edge of the input list heights (1st/last row and 1st/last column). By performing BFS we fill up each set with cells that are reachable. Using this, we get the intersection of both sets after the searches have concluded to find the common cells, because this means each cell can reach both oceans.



Loom Link
https://www.loom.com/share/d64c7a467e854f0ea1c5f8c86390bf84