Create 33_search_in_rotated_array.py
Solution Summary
The solution alters standard binary search to verify valid ranges of searching. We first check whether the pointers between "lo" and "mid" are in sorted order or not. In either event, we adjust pointers based on if the target is within range of "lo" to "mid" or "mid" to "hi". 

Loom Link
https://www.loom.com/share/33743a5b5e084b31b15103030c48d009