Create 62_unique_paths.py
Solution Summary
This function solves the unique paths problem by solving via bottom-up approach. We set the base cases of the last row and col to be all 1's. We then calculate the unique path of a given cell by the sum of the right neighbor and bottom neighbor. This will eventually lead us back to the top-left corner which will contain the number of unique paths from the top-left to bottom-right corner.

Loom Link
https://www.loom.com/share/d752d8c55cc243b4a7de310efaeaa650