Create 211_design_add_and_search_words_data_structure.py
Solution Summary
This solution uses the exact same setup from first Trie problem, in that we create TrieNodes of characters at each level when needed and mark the final character as a word. The only difference in the search method is the handling of the the "." character in which we must use a dfs function that keeps track of an idx and node. When we match with a "." we check if the next index and child lead to a word and return True/False based on this. Otherwise, we handle alphabetical characters the exact same in the first problem.

Loom Link
https://www.loom.com/share/abc58ff064c2430aa36edaea8f6c7a29