Create word_search_ii.py
Solution Summary
This solution uses the Trie data structure to store all the words we are looking for when we traverse the board. We can traverse the board using DFS, scanning our Trie at each step for valid words. We optimize the solution by adding references "refs", to our TrieNode's that give us the indication of whether we should continue searching or not (if ref < 1, we backtrack).

Loom Link
https://www.loom.com/share/3411a2d4fdcc467bbdf21210fbc155e1