Create coin_change.py
Solution Summary
The solution uses DP to calculate the minimum number of coins for each coin from 0 to the input amount. In doing so, we check each possible coin in the coins list for the minimum combination by calculating the difference of the current coin amount we are on (index i) - amount. Using this previously stored difference, we add one (because we are including the coin we are processing) to get the minimum value for the current index of dp[i].

Loom Link:
https://www.loom.com/share/076fdef896214abbbd561b452054ba5f