Create plus_one.py
Solution Summary
Iterate through digits in reverse order, adding one an returning when we find an index that doesn't have a value of 9. If we come across a 9, we replace it with a 0 and continue moving backwards. Should we fall off the list entirely, we will add a new index of value 1 of the front of digits.

Loom Link
https://www.loom.com/share/f4273e9c72754de8812dbd7524944a4c